refactor: split up `api/check.rs` file
